+++ bchain/src/config/config_utils.rs
@@ -1,13 +1,13 @@
 use crate::config::static_config;
 
 pub fn get_epoch(block_height: usize) -> usize {
-    block_height / static_config::EPOCH_HEIGHT
+    block_height / static_config::EPOCH_HEIGHT as usize
 }
 
 pub fn get_validators_consensus_block(epoch: usize) -> usize {
     if epoch < 2 {
         0
     } else {
-        (epoch - 1) * static_config::EPOCH_HEIGHT - 1
+        (epoch - 1) * static_config::EPOCH_HEIGHT as usize - 1
     }
 }
